+++ lite/utils/env.h
@@ -46,27 +46,6 @@
 #define QUANT_INPUT_OUTPUT_SCALE_RESTRICT_METHOD \
   "QUANT_INPUT_OUTPUT_SCALE_RESTRICT_METHOD"
 
-// The windows environmental variable min/max should be unset to avoid confilcts
-// with std::max/std::min
-#if !defined(_WIN32)
-#include <sys/time.h>
-#include <sys/types.h>
-#else
-#define NOMINMAX  // msvc max/min macro conflict with std::min/max
-#include <windows.h>
-#undef min
-#undef max
-extern struct timeval;
-static int gettimeofday(struct timeval* tp, void* tzp) {
-  LARGE_INTEGER now, freq;
-  QueryPerformanceCounter(&now);
-  QueryPerformanceFrequency(&freq);
-  tp->tv_sec = now.QuadPart / freq.QuadPart;
-  tp->tv_usec = (now.QuadPart % freq.QuadPart) * 1000000 / freq.QuadPart;
-  return (0);
-}
-#endif
-
 namespace paddle {
 namespace lite {
